requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
msal
python-dotenv